minor fix: toInt

+++ client/layout/src/LayoutManager.dart
@@ -103,7 +103,7 @@ class _LayoutManager extends RunOnceViewManager implements LayoutManager {
 			view.width = width();
 			break;
 		case _AmountInfo.RATIO:
-			view.width = (width() * amt.value).round();
+			view.width = (width() * amt.value).round().toInt();
 			break;
 		case _AmountInfo.CONTENT:
 			final int wd = view.measureWidth(mctx);
@@ -122,7 +122,7 @@ class _LayoutManager extends RunOnceViewManager implements LayoutManager {
 			view.height = height();
 			break;
 		case _AmountInfo.RATIO:
-			view.height = (height() * amt.value).round();
+			view.height = (height() * amt.value).round().toInt();
 			break;
 		case _AmountInfo.CONTENT:
 			final int hgh = view.measureHeight(mctx);
